+++ backend/src/services/dtos/Account/AccountInput.ts
@@ -3,4 +3,5 @@ import { AccountOutputDTO } from './AccountOutput';
 export type AccountCreateInputDTO = Pick<AccountOutputDTO, 'currency' | 'user_id'>;
 export type AccountUpdateInputDTO = Pick<AccountOutputDTO, 'balance' | 'id'>;
 
-export type AccountGetterDTO = AccountOutputDTO['id'] | AccountOutputDTO['user_id'];
+export type AccountGetAllInputDTO = AccountOutputDTO['user_id'];
+export type AccountGetterDTO = AccountOutputDTO['id'];
