+++ backend/src/utils/helpers.ts
@@ -5,6 +5,26 @@ const SYMBOLS = process.env.SYMBOLS as string;
 
 export const currencies: string[] = SYMBOLS.split(',');
 
+export const selectAccountOptions = { user_id: true, balance: true, currency: true, id: true };
+
+export const operators = {
+  equal: 'equals',
+  not: 'not',
+  gte: 'gte',
+  lte: 'lte',
+  in: 'in',
+};
+
+export const addFilters = (filters: { filterBy: any; value: any; operator: string }[]) => {
+  return filters.reduce((acumulator: { [key: string]: any }, current) => {
+    const operator = current.operator;
+    const key = current.filterBy;
+    const value = current.value;
+    acumulator[key] = { ...acumulator[key], [operator]: value };
+    return acumulator;
+  }, {});
+};
+
 export const errorStatusCodeMap = {
   VALIDATION_ERROR: 400,
   UNAUTHORIZED_ERROR: 401,
